Refactor run-local

+++ Makefile
@@ -1,18 +1,8 @@
 SHELL := /bin/bash
-.PHONY: help
 
+.PHONY: help
 help:
 	@echo -e "$$(grep -hE '^\S+:.*##' $(MAKEFILE_LIST) | sed -e 's/:.*##\s*/:/' -e 's/^\(.\+\):\(.*\)/\\x1b[36m\1\\x1b[m:\2/' | column -c2 -t -s :)"
-
-TAG	:= latest
-APP	:= nginx
-USER	:= library
-IMAGE	:= $(USER)/$(APP)
-
-run-docker: ## runs this app locally
-	docker run -it -d --rm\
-	   --name $(APP) \
-	   $(USER)/$(APP):$(TAG)
 
 include minikube-common.mk
 include kubectl-common.mk
@@ -20,16 +10,12 @@ include terraform-common.mk
 include docker-common.mk
 
 .PHONY: k8s-deploy
-k8s-deploy: ## Deploys this app on k8s
-	kubectl create -f ./manifests
+k8s-deploy: ## Deploys this app on k8s with kubectl
+	test -d manifests && kubectl create -f ./manifests
 
 .PHONY: k8s-undeploy
-k8s-undeploy: ## Undeploys this app from k8s
-	kubectl delete -f ./manifests
-
-.PHONY: k8s-run-app
-k8s-run-app: ## Runs this app on k8s
-	minikube service docker-homework-2-svc
+k8s-undeploy: ## Undeploys this app from k8s with kubectl
+	test -f manifests && kubectl delete -f ./manifests
 
 .PHONY: yamllint
 yamllint: ## Lint yaml manifests
